// packages/city-os-dashboard/src/libs/type.ts
export enum ConfigFormType {
  DEVICE_ONLY = 'DEVICE_ONLY',
  DEVICES_DURATION_LAYOUT = 'DEVICES_DURATION_LAYOUT',
  DEVICE_DURATION_LAYOUT = 'DEVICE_DURATION_LAYOUT',
  DIVISION_ONLY = 'DIVISION_ONLY',
  DIVISION_LAYOUT = 'DIVISION_LAYOUT',
  DEVICE_PLURAL_TITLE = 'DEVICE_PLURAL_TITLE',
  DEVICE_LAYOUT = 'DEVICE_LAYOUT',
  DEVICE_TEMPERATURE_UNIT_LAYOUT = 'DEVICE_TEMPERATURE_UNIT_LAYOUT',
  DEVICES_TITLE = 'DEVICES_TITLE',
  DEVICES_TITLE_LAYOUT = 'DEVICES_TITLE_LAYOUT',
}

export enum Duration {
  DAY = 'DAY',
  WEEK = 'WEEK',
}

export enum GadgetSize {
  DEFAULT = 'DEFAULT',
  SQUARE = 'SQUARE',
  RECTANGLE = 'RECTANGLE',
}

export type GadgetSizeType<T extends GadgetType> = typeof gadgetSizeSet[T][number];

export type GadgetConfigType<C extends ConfigFormType> = typeof configFormSet[C][number];

export interface GadgetConfig<ConfigType extends ConfigFormType> {
  id: string;
  width: number;
  height: number;
  type: GadgetConfigType<ConfigType>;
  setting: ConfigType extends ConfigFormType.DEVICE_ONLY
    ? {
        deviceId: string;
      }
    : ConfigType extends ConfigFormType.DEVICES_DURATION_LAYOUT
    ? {
        deviceIds: string[];
        duration: Duration;
        size: GadgetSize;
      }
    : ConfigType extends ConfigFormType.DEVICE_DURATION_LAYOUT
    ? {
        deviceId: string;
        duration: Duration;
        size: GadgetSize;
      }
    : ConfigType extends ConfigFormType.DIVISION_ONLY
    ? {
        groupId: string;
      }
    : ConfigType extends ConfigFormType.DIVISION_LAYOUT
    ? {
        groupId: string;
        size: GadgetSize;
      }
    : ConfigType extends ConfigFormType.DEVICE_PLURAL_TITLE
    ? {
        title: string;
        deviceIds: string[];
      }
    : ConfigType extends ConfigFormType.DEVICE_LAYOUT
    ? {
        deviceId: string;
        size: GadgetSize;
      }
    : ConfigType extends ConfigFormType.DEVICE_TEMPERATURE_UNIT_LAYOUT
    ? {
        deviceId: string;
        unit: TemperatureUnit;
        size: GadgetSize;
      }
    : ConfigType extends ConfigFormType.DEVICES_TITLE
    ? {
        title: string;
        deviceIds: string[];
      }
    : ConfigType extends ConfigFormType.DEVICES_TITLE_LAYOUT
    ? {
        title: string;
        deviceIds: string[];
        size: GadgetSize;
      }
    : never;
}

export enum TemperatureUnit {
  C = 'C',
  F = 'F',
}

export type GadgetsConfig = GadgetConfig<ConfigFormType>[];

export type ConfigSaveType = 'update' | 'create';

export interface BasicGadgetInfo {
  name: string;
  type: GadgetType;
  deviceType?: DeviceType;
  attribute?: AttributeInput;
  description: string;
  size: GadgetSize;
}

export interface DeviceOption {
  label: string;
  value: string;
}

export interface Curve {
  key: string;
  label?: string;
  points: Point[];
  variant?: 'curve' | 'areaClosed' | 'bar';
  color?: keyof Theme['palette']['gadget'];
}
// ...
export enum ExtremeOperation {
  MAX = 'MAX',
  MIN = 'MIN',
  SUM = 'SUM',
  COUNT = 'COUNT',
  AVG = 'AVG',
}